Extract statements_schema.go
